+++ RELEASENOTES.md
@@ -14,6 +14,9 @@ Thanks to qadan, cheuer, & compiling
 
 # Bug Fixes and Notes.
 
+* 0.4.0.8
+	* Ganon jokes added for when silvers aren't available
+	* Some text updated (Blind jokes, uncle text)
 * 0.4.0.7
 	* Reduce flashing option added
 	* Sprite author credit added